recent push fixed

+++ VisualPOVRAY/VisualPOVRAY/JuliaFractal.cs
@@ -67,7 +67,7 @@ namespace VisualPOVRAY
         }
 
 
-        public void update(int currentTime)
+        public void update(float currentTime)
         {
             if (reactive)
             {
