+++ VisualPOVRAY/VisualPOVRAY/Ovus.cs
@@ -48,7 +48,7 @@ namespace VisualPOVRAY
         }
 
 
-        public void update(int currentTime)
+        public void update(float currentTime)
         {
             if (reactive)
             {
